Added Battery Identification
Created battery identification, batteries will be numbered 1 to 416, and this can be accessed through the 'pack' matrix which contains the ordering of the batteries, and the 'pack2' matrix which contains the numerical identification of that battery. Currently this is just through numbering batteries 1-416, but preferably with Matlab2017 we can do A01-A30,B01-B30,C01-C30...N01-N26.