BatteryCell library is not useful library since it's functionallity can be replaced entirely with a simple array. Rm'ed. Added DAC functionallity to newBatteryTester and verified functionallity. Once channel can be used for I_SET, the other for setting the voltage reference on the ADC. MovedvDecoder pins over to D3---D7. Fixed a bug with the initialization of Decoder. (initial state of ENABLE) All SPI devices can now communicate on the same bus, as they should. (SD card reader required a 100 ohm resistor in series with MISO because MISO never goes high-Z like it should.) Discharger library abstracts away the DAC and PCB wiring so that setup and loop only deal with "cell" objects (arrays)
TODO:
The code in setup and loop dealing with myFile can be refactored into CSVwriter, since it's purpose was to abstract away some of the file handling.
Also additional code needs to be implemented to safely handle stopping and starting the discharger. I think doing this over the serial interface will be easiest (at least compared to implementing buttons and LCD display and interface)
Lastly, the safety checks should be implemented to prevent over-discharge of the cells and immediate cuttoff when the cells get too hot.